вынес handlers из формы создания/редактирования в отдельный файл, поправил пагинацию
