+++ src/app/taskForm/TaskForm.module.css
@@ -1,3 +1,9 @@
+.task-form {
+  display: flex;
+  flex-direction: column;
+  gap: 2em;
+}
+
 .header {
   font-size: 3rem;
 }
@@ -11,6 +17,23 @@
   color: white;
   padding: 2%;
   font-size: 1.5rem;
+  border: black;
+}
+
+.add-button:hover {
+  color: greenyellow
+}
+
+.input-container {
+  display: flex;
+  align-items: flex-start;
+  padding: 0;
+}
+
+.input {
+  font-weight: 600;
+  font-size: 1.5em;
+  padding: 1%;
 }
 
 .invalid-input {
