+++ src/app/taskList/TaskList.module.css
@@ -12,10 +12,11 @@
   color: white;
   font-weight: bold;
   font-size: 1.5rem;
+  margin-bottom: 10%;
 }
 
 .addButton:hover {
-  color: white
+  color: greenyellow;
 }
 
 .errorBlock {
